requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
multidict==6.0.5
openai==1.16.2
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
nanoid==2.0.0
